requirements.txt
streamlit>=1.30.0
folium>=0.15.0
streamlit-folium>=0.18.0
supabase>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
